Dialyzer: Fix all errors reported while testing testsuites
Some errors are due to incorrect specs, others are mistakes in the
testsuites themselves.

I had to disable some Dialyzer checks, sometimes for specific functions,
sometimes for an entire module. The reasons are that some functions are
incorrect on purpose to trigger an exception or a demonstrate a misuse,
and some functions call unexported functions (like non-existing
functions or functions exported when compiled with `-DTEST` only).

While working on this patch, an unsupported Beam instruction was
discovered: `{test, _, Label, _, _}`. Support is added in this patch
too.