khepri_fun: Support binary handling instructions
In the process, we discovered the special `0' label. This is used to
trigger an exception if the instruction using this label jumps to that
special label.

Also, we had to handle the decoding of the `field_flags` argument to
some of those binary instructions. Indeed, the decoding is not yet
performing it when we ask `beam_disasm` to disassemble a module (it's
still the integer bit fields, not a list of atoms as expected by the
compiler).