khepri_fun: Reject extraction of unexported external functions
An external unexported function is `fun module:function/0` when `module`
has `function/0` but doesn't export it.

We can't allow to extract this type of function because its assembly
code might not work outside of the context it was compiled for. For
instance, it could have some optimizations which make Erlang crash
because they only work in the original module.

Without this, extracting and executing `inner_function/2` in the
testsuite added in the previous commit woudl cause Erlang to termine
with a Segmentation fault.