khepri_fun: Add guessed `var_info` annotations to assembly
Let's take the following code:

    -export([outer_function/2]).

    outer_function(#my_record{field = ...} = MyRecord, Term) ->
        (...),
        inner_function(MyRecord, Term);
    outer_function(_, _) ->
        (...).

    inner_function(#my_record{field = ...} = MyRecord, Term) ->
        (...);
    inner_function(_, _) ->
        (...).

Both `outer_function/2` and `inner_function/2` take a tuple as their
first argument and do a pattern matching with it. `outer_function/2` is
exported and `inner_function/2` is only called by `outer_function/2`.

In this situation, the compiler may perform an optimization on
`inner_function/2`: it is sure that the function will take a tuple of
the expected form because it was already checked in `outer_function/2`
and it is the only caller. Therefore, it will skip any verifications and
will read the content of the tuple directly using the
`get_tuple_element` instruction.

When compiled from source, the compiler adds annotations of the
following form at the beginning of `inner_function/2` assembly:

    {'%', {var_info, {x,0}, [{type, {t_tuple, 2, true, #{...}}}]}}.

However, the assembly we recover from an already compiled module misses
those annotations. And when we compile the same assembly without those
annotations, the compilation fails with:

    {bad_type,
     {needed, {t_tuple, _Size, _, _Fields} = NeededType},
     {actual, any}}

This is caused by checks that we are only passing tuples to a
`get_tuple_element` instruction for instance.

This patch handles this compilation failure by forging a `var_info`
annotation based on the error message and adding it at the beginning of
the failing function. The compilation is then retried.

The recreated annotation doesn't exactly match the original one. But it
is sufficient to please the compiler. I make the assumption this is fine
because the code we extract was already compiled successfully once.
Therefore, there should be no situation where for example
`get_tuple_element` would be called with something else than a tuple.

If the annotated code still fails to compile with a similar error for
the same variable/register annotation, the extraction aborts. This is to
prevent infinite retry loops because of an incorrectly forged `var_info`.

This added error handling matches specifically the situation described
above for now. We may add more in the future based on real situations we
encounter.