khepri_fun: Patch incorrectly decoded instructions
`beam_asm` decodes several bitstring matching instructions incorrectly.
Sometimes, it's just the field flags which are not decoded.
Unfortunately, sometimes, instruction arguments order is incorrect or
the argument does not have the form expected by the compiler.

Now, `pass1_process_instructions/3` starts by patching those
instructions before doing anything else with them. This way the rest of
the code handles correct instructions (from the point of view of the
compiler). A second benefit is that if `beam_disasm` is fixed, Khepri
will continue to work.