add translation for 'raise' BIF
See this line in beam_asm:

    https://github.com/erlang/otp/blob/df48c260e74c3e9058ff8681ce9f554e6fa0fe34/lib/compiler/src/beam_asm.erl#L360-L361

The `raise` instruction needs to be translated to a `bif`. Luckily
the arguments line up very well so it's an easy translation. It was
a bit tough writing a test case that the compiler wouldn't optimize
the raise/try out of, I ended up using the send+receive technique from
a few test cases down.