use instructions as locations when inserting annotations
add_comment_to_code/4 now takes Location as a pair

    {Placement :: before | 'after', Instruction :: tuple() | atom()}

And will use the instruction as the insertion point, placing the
annotation before or after it. This resolves some awkwardness from
using Location as an (1-indexed) index in the list of instructions of
a function. The instruction index works but is brittle when hard-coded
as it was for the `no_bs_start_match2' instruction. For example if
line instructions were not discarded, it could be impossible to tell
where in the instruction list the annotation should go.