khepri_fun: Provide meaningful `line` instructions
Line information comes from the "Line" chunk in the beam module. This
is not decoded by `beam_disasm` and the `line` instructions it produces
only have the index inside that "Line" beam chunk.

So far, we needed to get rid of all those `line` instructions because
they were invalid for the compiler. It meant that frames in stacktraces
pointing to an extracted function had no location. This made any sort of
debugging quite difficult.

Here is an example without this patch:

    %% Based on the following function:
    %%     crashing_fun() ->
    %%         throw("Expected crash").

    1> Fun = khepri_fun:to_standalone_fun(fun mod:crashing_fun/0, #{}).
    2> khepri_fun:exec(Fun, []).
    ** exception throw: "Expected crash"
         in function  kfun__mod__crashing_fun__33048370:run/0

Now, we decode the "Line" beam chunk for each module we need to
disassemble. We then use this decoded table to replace the incomplete
`line` instructions by valid ones.

Thanks to that, we get stacktrace frames with correct locations, making
debugging far easier. Here is the same example as above, but with this
change in place:

    1> Fun = khepri_fun:to_standalone_fun(fun mod:crashing_fun/0, #{}).
    2> khepri_fun:exec(Fun, []).
    ** exception throw: "Expected crash"
         in function  kfun__mod__crashing_fun__33048370:run/0 (.../mod.erl, line 30)

Fixes #50.