discard existing comments when merging annotations
Here we peek at the next instruction in either Result or Rest
(depending on Placement) to find an existing comment if there is
one and then we discard the existing comment after merging.
`beam_validator' currently accepts duplicate annotations for a
types: it discards all but the last annotation before a variable
is checked. For example if we were fixing the typing of a
`bs_start_match4`, we might end up with:

    {'%', {var_info, {x, 0}, [accepts_match_context]}},
    {'%', {var_info, {x, 0}, [{type, {t_bs_context,1,0,0}},
                              accepts_match_context]}},
    {bs_start_match4,{atom,resume},2,{x,0},{x,0}}.

Currently `beam_validator` behaves as though it discards the first
annotation. This behavior is not documented, though, so we protect
against a future where `beam_validator' stops handling duplicates
like this by ourselves discarding the existing annotation.