re-throw compilation failure when an annotation already exists
Here we avoid an infinite loop of compilation that can happen
when we attempt to fix a compilation error by adding an annotation
and that annotation already exists where we expect. This catches the
case where we expect an annotation to fix a compilation error, we
add the annotation, and then the compiler fails with the same error.