khepri_fun: Work around bad decoding of `get_map_elements` arguments
The last argument of `get_map_elements` is a list of (possibly tagged)
registers. The tagged registers are incorrectly decoded by `beam_disasm`
and we must patch them like we do it already in other places.

This fixes an error where a standalone function is successfully compiled
but fails to load.