khepri_fun: Work around bad decoding of `{test, is_tuple, ...}` arguments
Likewise for `{test, is_tagged_tuple, ...}`.

The last argument is a list of (possibly tagged) registers. The tagged
registers are incorrectly decoded by `beam_disasm` and we must patch
them like we do it already in other places.

This fixes an error where a standalone function is successfully compiled
but fails to load.