khepri_fun: Patch `{test, ...}` in pass 1's third group
The first group detects bad decoding from the function clause, patches
the instruction and call `pass1_process_instructions()` again on the
same instruction.

We can't detect all those errors from a function clause, that's why we
patch those instructions in the third group, at the same time we add
`var_info` type information for instance.

This commit moves the fix to `{test, ...}` to the third group to
maintain the logic of each group.