khepri_fun: Handle FunIndex variant of call_fun2/3
call_fun2/3 has three possible tags for the second element in the
instruction tuple:

    ## @spec call_fun2 Tag Arity Func
    ## @doc Calls the fun Func with arity Arity. Assume arguments in registers x(0)
    ##      to x(Arity-1). Tag can be one of:
    ##
    ##      * FunIndex      - `Func` is always a local fun identified by `FunIndex`
    ##      * {atom,safe}   - `Func` is known to be a fun of correct arity.
    ##      * {atom,unsafe} - Nothing is known about `Func`.
    178: call_fun2/3

We already handle the {atom,safe} variant. By nesting some higher
order functions and using a local function as one of the called
functions, we can produce the FunIndex variant.

When translated correctly, this variant has a {f,Label} shape
where the Label is known to be the beginning of the called local
function's scope.

    {call_fun2,
     {f,821},
     3,
     {tr,{x,3}, {t_fun,3,{'-projection_fun_for_sets/1-fun-1-',4},any}}}.

But beam_disasm gives the shape:

    {call_fun2,
     {u,57},
     3,
     {tr, {x,3}, {{t_fun,any,any,any},0,18446744073709551615}}.

The type-tagged register must be decoded to discard the surrounding
information similar to other instructions that use type-tagged
registers.

The {u,?} register doesn't point to a recognizable label index so
we can't determine the correct function index to use. However we
know that if the compiler emitted the FunIndex variant of
call_fun2/3, it must know that a local function of the correct arity
is being called. So we can replace the {u,?} tag with {atom,safe}
which signals that the function is known to be the correct arity.