Transactions: Support stored procedures as transaction functions
This allows to extract and store a transaction function once and use it
many times. This saves the extraction time and reduce the size of the
transaction Ra command.

Here is an example:

    generic_tx(Wood, Increase) ->
        Path = [stock, Wood],
        {ok, Stock} = khepri_tx:get(Path),
        NewStock = erlang:min(?MAX_STOCK, Stock + Increase),
        ok = khepri_tx:put(Path, NewStock).

    store_wood_stock_tx() ->
        khepri:put(?STORE_ID, [tx, wood_stock], fun generic_tx/2).

    increase_wood_stock(Wood, Increase) ->
        khepri:transaction(
          ?STORE_ID, [tx, wood_stock], [Wood, Increase]).

The constraints for transaction functions remain the same: they can't
have side effects. So now, everytime a function is stored in Khepri, we
verify during extraction if it can be used as a transaction function and
store the result with the extracted function. When the stored procedure
is executed as a transaction function, we check the result of that
verification and possibly deny the execution.

Fixes #141.