khepri_fun: Fix translation for bs_get_float2
The resolution for `bs_get_float2` in `beam_disasm` is in the same
block and has the same body as the translations for `bs_get_integer2`
and `bs_get_binary2` so it has the same mistranslations and can be
fixed in the same way.

We can provoke a `{test,bs_get_float2,..}` instruction by pattern
matching on a bitstring that uses the `float` type specifier.