New library created from the `khepri_fun` module in Khepri
We wanted to extract the `khepri_fun` module and make an independent
library out of it because the functionality is useful on its own.
Another benefit is that it will be easier to maintain, test and
document.

The new name is Horus, to continue on this trend of using names from the
Egyptian mythology.