Cleanup headers + expose helpful macros
[Why]
It can be useful to check if an argument is a standalone function or
what its arity is right from a function clause instead of a function
body.

[How]
We introduce the `include/horus.hrl` header which contains:
* macros to identify and query a #standalone_fun{} record without
  exposing the record itself;
* macros to match errors and exceptions raised by Horus.

Note that the #standalone_fun{} header remains internal and opaque to
users of Horus. This is to prevent the need to recompile applications
consuming this record whenever we change it.