Fix cached standalone fun used in fun environment
[Why]
If an anonymous function uses another anonymous function in its
environment and this other function was already processed (i.e. it's in
the standalone functions cache), then we lack the `all_calls` map which
is used at the end of the whole extraction to determine if the top-level
processed function is still needed.

Without this `all_calls` map, the `is_standalone_fun_still_needed/1`
works on incorrect inputs. The result is that it may incorrectly decide
that a standalone function is not needed because it didn't know that a
function in the environment performed some calls.

[How]
We simply need to retain the `all_calls` map in the cache, like the
checksums and the processing options. This way, we use this map in the
processing of the environment and the top-level function. This fixes the
bug because `is_standalone_fun_still_needed/1` now has the correct
inputs.